Add cacheobject.UsingRequestResponseWithObject API
It's nice to not have to rewrite all of UsingRequestResponse just because you
want to check out one random field on obj (say, to look for
NoCache/NoCachePresent if you don't have the ability to revalidate).

It's also nice to have access to the OutWarnings field, so I added that to this
advanced API as well.